+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Changed 
 
+- Small bug fix for `validity` function.
 - Improved `find_potential_neighbours` method.
 - Added possibility to run `compute_divergence` for a fixed number of samples.